# phase-3/specs/001-backend-api/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Todo Backend API
  description: RESTful API for task management with CRUD operations and database persistence
  version: 1.0.0
  contact:
    name: Backend API Development Team

servers:
  - url: http://localhost:8000
    description: Local development server

tags:
  - name: tasks
    description: Task CRUD operations
  - name: health
    description: Health check endpoints

paths:
  /api/tasks:
    get:
      tags:
        - tasks
      summary: List all tasks
      description: Retrieve all tasks in the system
      operationId: listTasks
      responses:
        '200':
          description: Successful response with list of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskResponse'
              example:
                - id: "550e8400-e29b-41d4-a716-446655440000"
                  user_id: "660e8400-e29b-41d4-a716-446655440001"
                  title: "Complete project documentation"
                  description: "Write comprehensive API documentation"
                  due_date: "2026-02-15T17:00:00Z"
                  completed: false
                  created_at: "2026-02-08T10:00:00Z"
                  updated_at: "2026-02-08T10:00:00Z"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - tasks
      summary: Create new task
      description: Create a new task with title and optional description/due date
      operationId: createTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
            example:
              title: "Complete project documentation"
              description: "Write comprehensive API documentation"
              due_date: "2026-02-15T17:00:00Z"
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          description: Bad request - validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                detail: "Title cannot be empty"
        '422':
          description: Unprocessable entity - validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/tasks/{task_id}:
    get:
      tags:
        - tasks
      summary: Get single task
      description: Retrieve a specific task by its ID
      operationId: getTask
      parameters:
        - name: task_id
          in: path
          required: true
          description: UUID of the task to retrieve
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful response with task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                detail: "Task not found"
        '422':
          description: Invalid UUID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - tasks
      summary: Update task
      description: Update all fields of an existing task
      operationId: updateTask
      parameters:
        - name: task_id
          in: path
          required: true
          description: UUID of the task to update
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
            example:
              title: "Updated task title"
              description: "Updated description"
              due_date: "2026-02-20T17:00:00Z"
              completed: true
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          description: Bad request - validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                detail: "Task not found"
        '422':
          description: Unprocessable entity - validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - tasks
      summary: Delete task
      description: Delete a task by its ID
      operationId: deleteTask
      parameters:
        - name: task_id
          in: path
          required: true
          description: UUID of the task to delete
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Task deleted successfully (no content)
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                detail: "Task not found"
        '422':
          description: Invalid UUID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/tasks/{task_id}/complete:
    patch:
      tags:
        - tasks
      summary: Toggle task completion
      description: Toggle the completion status of a task (completed ↔ incomplete)
      operationId: toggleTaskCompletion
      parameters:
        - name: task_id
          in: path
          required: true
          description: UUID of the task to toggle
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Task completion status toggled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                detail: "Task not found"
        '422':
          description: Invalid UUID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /health:
    get:
      tags:
        - health
      summary: Health check
      description: Check if the API is running and database is accessible
      operationId: healthCheck
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "healthy"
                  database:
                    type: string
                    example: "connected"
        '503':
          description: Service unavailable
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "unhealthy"
                  database:
                    type: string
                    example: "disconnected"

components:
  schemas:
    CreateTaskRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          description: Task title (required, cannot be empty)
          example: "Complete project documentation"
        description:
          type: string
          nullable: true
          description: Optional detailed description
          example: "Write comprehensive API documentation"
        due_date:
          type: string
          format: date-time
          nullable: true
          description: Optional due date in ISO 8601 format
          example: "2026-02-15T17:00:00Z"

    UpdateTaskRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          nullable: true
          description: Updated task title
          example: "Updated task title"
        description:
          type: string
          nullable: true
          description: Updated task description
          example: "Updated description"
        due_date:
          type: string
          format: date-time
          nullable: true
          description: Updated due date in ISO 8601 format
          example: "2026-02-20T17:00:00Z"
        completed:
          type: boolean
          nullable: true
          description: Updated completion status
          example: true

    TaskResponse:
      type: object
      required:
        - id
        - user_id
        - title
        - completed
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
          description: Unique task identifier
          example: "550e8400-e29b-41d4-a716-446655440000"
        user_id:
          type: string
          format: uuid
          description: User who owns this task
          example: "660e8400-e29b-41d4-a716-446655440001"
        title:
          type: string
          description: Task title
          example: "Complete project documentation"
        description:
          type: string
          nullable: true
          description: Task description
          example: "Write comprehensive API documentation"
        due_date:
          type: string
          format: date-time
          nullable: true
          description: Task due date in ISO 8601 format
          example: "2026-02-15T17:00:00Z"
        completed:
          type: boolean
          description: Completion status
          example: false
        created_at:
          type: string
          format: date-time
          description: Timestamp when task was created
          example: "2026-02-08T10:00:00Z"
        updated_at:
          type: string
          format: date-time
          description: Timestamp when task was last updated
          example: "2026-02-08T10:00:00Z"

    ErrorResponse:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          description: Error message
          example: "Task not found"

    ValidationErrorResponse:
      type: object
      required:
        - detail
      properties:
        detail:
          type: array
          items:
            type: object
            properties:
              loc:
                type: array
                items:
                  type: string
                description: Location of the validation error
              msg:
                type: string
                description: Error message
              type:
                type: string
                description: Error type
          example:
            - loc: ["body", "title"]
              msg: "field required"
              type: "value_error.missing"